+++ src/org/lukasj/idea/torquescript/psi/TslElementType.kt
@@ -1,0 +1,8 @@
+package org.lukasj.idea.torquescript.psi
+
+import com.intellij.psi.tree.IElementType
+import org.jetbrains.annotations.NonNls
+import org.lukasj.idea.torquescript.TslLanguage
+
+class TslElementType(debugName: @NonNls String) : IElementType(debugName, TslLanguage.INSTANCE) {
+}